Update buttons to be more visible and cover the hover bug on mobile
On mobile, the :hover state is quite problematic as we don't know whether the hover is oover or not, therefore cause sometimes the hover to be stuck.

By changing the colors to a constant filled, the hover is less noticeable, and therefore, if the hover problem happens, it will be way less visible.