Tweak loader initializator to accept divs root loaders
Also changed homepage's loader to be a div instead of a form.

This change also uses the files checker function for the input click instead of drag & drop, meaning the file loading portion does not need to check the files too, removing code duplicatas.

The change in accepting divs instead of forms only allows for file loaders insides of form, because you cannot nest forms in HTML.

Ref #60