Tweak image previewer layout to normalize whitespace
This commit makes the whitespace consistent throught different screen sizes by removing the col aspect f the layouting structure.

I have changed the code now so that the field and the remover button are always in the same row, with a constant whitespace (margin), and a previewer that always takes the most space it can.