Add button and input field to homepage
No logic is yet done to grab the code and nothing else really is done - just the custom input field.

This took a while to figure out because the mobile events are not triggered by the keydown events in Android (Chrome) - I will probably need to test this custom input field with different brosers soon...

This custom input field also currently does not support pasting, which is something I wish to add sooner or later.