Add logic to send data to API on create shared election button press
This commit also makes the images compressed in the data object - useful to not overload my server. I also plan to set a limit of 2 MB for the image, otherwise my server might be given gigabites of images through my actual webpage, without modifications. This limit will also be imposed in my server for security mesures.

The data "isDownloadDisabled" was also added to the data object to make it go through a shared election.

There is not yet a modal to confirm if the data can be sent to my server - I'll tackle this in a later commit.