Tweak vote updating in shared election to respect other devices data
Few more requests are sent to update the data to match what other devices do.

This commit also allows that whenever you press space (the keyboard shortcut to show the toast notification in the votig screen), an update is requested and the current data is reflected (so doing votes on another machine makes the first device "show" progress in the toast notification).

This commit also transforms all request into async / awaited requests, as I'm always waiting for the response to continue a generic aproach that matches the non-shared version.