Implement new sendRequest method to check if server accessible
This new sendRequest method is also used to now send a request when trying to join a shared election and the erreors are handled in the viewlet.

The checks for the home's join shared election button shows as green if available, and red if not accessible, with a popover explaining the icon's meaning.

The join election button in the join shared election modal is now disabled until 6 chars are inputted in the inputs, and disabled while the requuest in on-going. This allows a spinner to show, and once the request is done, both the spinner and the disabled state are reset to their corresponding states. This same behavior is done to the partionned inputs, which gets all disabled for the duration of the request.