Add buttons to delete individual candidates and switch to remove all
This allows to remove directly the candidate you want to remove, and allows to also remove all of them if that is what is desired.

The button to remove all requires a confirmation, done through a popover that shows up on top of the remove all candidates button.

The buttons HTML elements are now whitelisted in bootstrap, so every tooltip / popovers will be able to render buttons if HTML is enabled for it.

Closes #67