Add new sendRequestFor method to send multiple request if failures
This new method send a request multiple times if it fails to ensure no mid-request errors happens.

The number of attempts is determined by the new parameter, and the underlying structure uses my own sendRequest method.

Finally, this method is async to aloow for waiting until the request is done to check if there is errors or not : it therefore returns a Promise that may fail, giving the exception with the request error details in it.