Add error handling for all voting requests
This error handling happens only when trying to get to the next voter if the votes send request failed, as the kids may see the error and this is unwanted.

The same modal is shown for all errors for this voting process, with some details changed to reflect what the error is about.

Whenever there is an error, the choice to retry or continue with a local election is given, allowing the admins to better handle how the election flow is going in case there is a general internet shortage.

The loading animation is in the top right of the page and again, is only shown when making a request when going to the next voter.