Tweak sendRequest method to add a minimum delay before ending request
This delay helps allowing the end user to see the spinner animation for a specified delay (which defaults to 500 milliseconds).

This also prevents jumps when the request is too fast because the spinner shows and hides immediately.

The only drawback with this is that now the usage of sendRequest requires the use of a catch, or an Uncaught error in a promise would be shown in the console. This catch clause could be empty if no handling needs to be done, but is now required.