Update pre-voting / voting handlers to better abstract flow logic
This commit abstracts the logic to go to the next voter to make it more generalized : it can now be used to start the election too (starting the elections is technically going to the next voter).

This allows better promises handling and prevents few code duplicatas.

This commit also adds handling for pre-voting requests errors, allowing you to start the elections locally if there was an error during the start of the downloaded shared election.