Handle skip request errors directly in toast
This commit also tweaks the CSS to fix a small z-index issue (popup wouldn't show above the toast because the bootstrap popover z-index was smaller than my custom z-index)

There is no confirmation to skip the election to a local instead, so this is to keep in mind.

I also just realized that skipping locally implies that there is a seat that will be taken but not used in the server, which might be problematic until handling leavers isn't done yet. To my todo list!

The toast now stays activated when pressing space multiple times, which simply refreshes the timeout before it closes itself. When the focus is on the confirm popover, the toast will not close itself until the focus is out of the popover or the error is handled.