Fix few incoherent behaviors and improve texts based on context
The toast notification in voting now shows a better text if the data was found to be skipped.

The post shared results shouldn't show an error for the verify button if the data was skipped and everyone voted correctly. That one was weird, I dunno why I added it yesterday, gotta love copy pasting few stuff...

Also fix undeclared variable usage