Fix flex elements not being hidden in updateSharedElectionCode method
When running the updateSharedElectionCode method, all shared election containers (and the "not containers") would get hidden based on wheter the code is truthy or not. However, `display: flex` elements does not ablidge to the `hidden` attribute, and were therefore not hidden correctly.

This commit fixes that (most noticeable were on the voting toast, with double buttons...)