Tweak setup input verifications request to only verify after delay
This delay between user inputs allows to quickly type the name of a candidate, for example, without the request to be sent multiple times while it is being loaded.

Also, the visual is now properly updated when an inputs gets unvalidated, rather than still showing the loading image and showing the good or bad icon.

Ref #75