Tweak handling of election data through its own class
This class lets me be way more flexible, as well as defining a general place to mess with the wanted data.

If I need to make a drastic change, I can just make it here, no worries! Actually, I even already added a backward compatibility feature in here, as I previously did it in the voting logic, which was working but unwanted.

The databases will now always be sanitized through this new class.