Improve input partitionner script by adding length attribute
When you add the `data-length` attribute to the input partitionner root, the script will init your input partition with the provided length, copying the first input node that it finds in the child nodes of your root. Therefore, you simply need to define one input type, and it will be copied the number of times you decided in your attribute.

Another upgrade is there to automatically set the `maxlength` attribute to 1, as this kind of inputs requires to be only 1 in length.