Add URL code handling to automatically open shared election prompt
When the URL contains a `?code=XXXXXX` query, the index script will automatically open the prompt to join a shared election, prepopulating the code inputs, to make sharing an election faster.

Currently, handling if the code is present in the query but the server is not accessible. I will add a toast showing that the server is not accessible if query code present but server not accessible.