Update voting's timeout scrolling fix to container option
Instead of relying on a timeout that cannot exactly be trusted, we now use the new requesterContainerOptions system of `Utils.sendRequest` to execute the scrollIntoView function right after the container was properly shown.

Just for a heads up, the previous 30 ms of timeout was to let the requester container show up. Yikes.